refactor(overview): replace memoized chart color mapping with direct reference; utilize custom hook for pie data handling
